Updated routes, controller path, test files, and README to use versioning in the url.
